+++ familyhelper-ui-view/packages/webapp-mobile/src/components/comvisual/widget/floatTool/types.ts
@@ -32,3 +32,23 @@ export type DockStatus =
   | DockStatusLeft
   | DockStatusBottom
   | DockStatusRight
+
+/**
+ * 调整状态：未调整。
+ */
+export type AdjustStatusIdle = 0
+
+/**
+ * 调整状态：正在鼠标按下。
+ */
+export type AdjustStatusPointerDown = 1
+
+/**
+ * 调整状态：正在移动。
+ */
+export type AdjustStatusMoving = 2
+
+/**
+ * 调整状态。
+ */
+export type AdjustStatus = AdjustStatusIdle | AdjustStatusPointerDown | AdjustStatusMoving